Spinning hand crank added to the side of the box.
Added a spinning hand crank that spins with the box, but with a different x-axis rotation.

Also adjusted how the top face of the cube is formed and created a new panel object. Currently, these changes are not visible in the animation, and are not functional. Future work would include making the top panel hinge open.